requirements.txt
html2text
pyyaml